[Cleanup] Replace displayError with printf, optimize and cleanup code
v0.3.5

+++ Core/Effects/effectsChain.cpp
@@ -9,31 +9,31 @@ EffectsChain::EffectsChain() {
 }
 
 EffectsChain::~EffectsChain() {
-    for (int i = 0; i < 4; ++i) {
+    for (uint8_t i = 0; i < 4; ++i) {
         delete pedals[i];
     }
 }
 
 void EffectsChain::clear() {
-    for (int i = 0; i < 4; ++i) {
+    for (uint8_t i = 0; i < 4; ++i) {
         delete pedals[i];
         pedals[i] = Pedal::createPedal(PedalType::PASS_THROUGH);
     }
 }
 
-void EffectsChain::setPedal(int index, PedalType type) {
+void EffectsChain::setPedal(uint8_t index, PedalType type) {
     if (index < 0 || index >= 4) return;
 
     delete pedals[index]; 
     pedals[index] = Pedal::createPedal(type);
 }
 
-Pedal* EffectsChain::getPedal(int index) const {
+Pedal* EffectsChain::getPedal(uint8_t index) const {
     return (index >= 0 && index < 4) ? pedals[index] : nullptr;
 }
 
 void EffectsChain::draw() const {
-    for (size_t i = 0; i < 4; ++i) {
+    for (uint8_t i = 0; i < 4; ++i) {
         if (pedals[i]) {
             const Bitmap& bmp = pedals[i]->getImage();
             Display::drawBitmap(bmp, pedalPositionsX[i], PEDAL_PAGE_START);
